Changed calculation for `threadCount` to prevent unsigned underflow. (#287)
* Changed calculation for `threadCount` to prevent unsigned underflow.

* Change several things to int32_t.

+++ src/resource/ResourceManager.cpp
@@ -15,14 +15,14 @@ extern bool SFileCheckWildCard(const char* szString, const char* szWildCard);
 namespace LUS {
 
 ResourceManager::ResourceManager(const std::string& mainPath, const std::string& patchesPath,
-                                 const std::unordered_set<uint32_t>& validHashes, uint32_t reservedThreadCount) {
+                                 const std::unordered_set<uint32_t>& validHashes, int32_t reservedThreadCount) {
     mResourceLoader = std::make_shared<ResourceLoader>();
     mArchive = std::make_shared<Archive>(mainPath, patchesPath, validHashes, false);
 #if defined(__SWITCH__) || defined(__WIIU__)
     size_t threadCount = 1;
 #else
     // the extra `- 1` is because we reserve an extra thread for spdlog
-    size_t threadCount = std::max(1U, (std::thread::hardware_concurrency() - reservedThreadCount - 1));
+    size_t threadCount = std::max(1, (int32_t)(std::thread::hardware_concurrency() - reservedThreadCount - 1));
 #endif
     mThreadPool = std::make_shared<BS::thread_pool>(threadCount);
 
@@ -33,14 +33,14 @@ ResourceManager::ResourceManager(const std::string& mainPath, const std::string&
 }
 
 ResourceManager::ResourceManager(const std::vector<std::string>& otrFiles,
-                                 const std::unordered_set<uint32_t>& validHashes, uint32_t reservedThreadCount) {
+                                 const std::unordered_set<uint32_t>& validHashes, int32_t reservedThreadCount) {
     mResourceLoader = std::make_shared<ResourceLoader>();
     mArchive = std::make_shared<Archive>(otrFiles, validHashes, false);
 #if defined(__SWITCH__) || defined(__WIIU__)
     size_t threadCount = 1;
 #else
     // the extra `- 1` is because we reserve an extra thread for spdlog
-    size_t threadCount = std::max(1U, (std::thread::hardware_concurrency() - reservedThreadCount - 1));
+    size_t threadCount = std::max(1, (int32_t)(std::thread::hardware_concurrency() - reservedThreadCount - 1));
 #endif
     mThreadPool = std::make_shared<BS::thread_pool>(threadCount);
